refactor: Colocate opengraph image gen code.
