+++ app/blog/[id]/opengraph-image.tsx
@@ -1,15 +1,12 @@
-import { ImageResponse } from 'next/og';
-import { readFile } from 'node:fs/promises';
-import { join } from 'node:path';
+import {
+    contentType,
+    generateBlogPostImage,
+    size,
+} from '../../../og/generateOpenGraphImage';
 import { getPostMetadata } from '../../../utils/getPostMetadata';
 import { Params } from './types';
 
-export const size = {
-    width: 1200,
-    height: 630,
-};
-
-export const contentType = 'image/png';
+export { contentType, size };
 
 type Props = {
     params: Promise<Params>;
@@ -18,83 +15,5 @@ type Props = {
 export default async function Image({ params }: Props) {
     const { id } = await params;
     const post = await getPostMetadata(id);
-
-    const karrikRegular = await readFile(
-        join(process.cwd(), 'assets/opengraph/Karrik-Regular.ttf')
-    );
-
-    const portraitData = await readFile(
-        join(process.cwd(), 'assets/opengraph/portrait.jpg')
-    );
-    const portraitSrc = Uint8Array.from(portraitData).buffer;
-
-    return new ImageResponse(
-        (
-            <div
-                style={{
-                    width: '100%',
-                    height: '100%',
-                    background: '#FFF',
-                    display: 'flex',
-                    flexDirection: 'column',
-                }}
-            >
-                <div
-                    style={{
-                        flexGrow: 1,
-                        padding: '32px 56px',
-                        color: '#000',
-                        fontSize: 96,
-                    }}
-                >
-                    {post.title}
-                </div>
-                <div
-                    style={{
-                        flexGrow: 0,
-                        flexShrink: 1,
-                        display: 'flex',
-                        fontSize: 56,
-                        justifyContent: 'space-between',
-                        alignItems: 'center',
-                        paddingLeft: '32px',
-                        paddingRight: '32px',
-                        paddingBottom: '32px',
-                        color: '#212224',
-                    }}
-                >
-                    <span
-                        style={{
-                            display: 'flex',
-                            alignItems: 'center',
-                            gap: '18px',
-                        }}
-                    >
-                        {/* eslint-disable-next-line @next/next/no-img-element, jsx-a11y/alt-text */}
-                        <img
-                            // @ts-expect-error I'm just doing what the docs say
-                            src={portraitSrc}
-                            style={{
-                                borderRadius: '50%',
-                                height: 100,
-                            }}
-                        />
-                        <span>Zach Snoek</span>
-                    </span>
-                    <span>zachsnoek.com</span>
-                </div>
-            </div>
-        ),
-        {
-            ...size,
-            fonts: [
-                {
-                    name: 'Karrik',
-                    data: karrikRegular,
-                    style: 'normal',
-                    weight: 400,
-                },
-            ],
-        }
-    );
+    return generateBlogPostImage({ title: post.title });
 }
